разбивает код на модули, добавляет приблизительную файловую структуру скриптов
