+++ js/util.js
@@ -1,0 +1,8 @@
+const getRandomInteger = function(min, max) {
+  const rand = min + Math.random() * (max + 1 - min);
+  return Math.floor(rand);
+};
+
+const getRandomArrayElement = (array) => array[getRandomInteger(0, array.length - 1)];
+
+export { getRandomInteger, getRandomArrayElement };
